Update CapsuleProgressView.swift animation style
The animation style for CapsuleProgressView.swift was changed from .easeOut to .linear. This change will modify the animation behavior of the CapsuleProgressView, resulting in a linear animation style.

Remove unused import

The import statement for SDWebImageSwiftUI in StepsView.swift was removed as it was not being used.

Make progress navigation gestures respond to drag

The progress navigation gestures in StepsView.swift now respond to drag gestures as well as taps. Dragging to the left will advance the progress by -1, while dragging to the right will advance the progress by 1.

Update ContentView.swift file reference

The file reference for ContentView.swift in project.pbxproj was updated to 1B18D7024AE807CA4D546173 from 1B18D1D177DD2E44FA853AE7. This change ensures that the correct file is referenced in the project.